feat: use write key
