+++ packages/contentful--create-contentful-app/src/analytics.ts
@@ -1,7 +1,7 @@
 import Analytics from 'analytics-node';
 
 // Public write key scoped to data source
-const SEGMENT_WRITE_KEY = 'define me'
+const SEGMENT_WRITE_KEY = 'rU9puFk7E1NlAkD0UXCdy6tEFHXc2OyN'
 
 interface CCAEventProperties {
   template?: string; // can be example, source or JS or TS
@@ -17,9 +17,14 @@ export function track(properties: CCAEventProperties) {
   const client = new Analytics(SEGMENT_WRITE_KEY);
 
   try {
-    client.track({event: 'app-cli-cca-creation', properties, timestamp: new Date() });
+    client.track({
+      event: 'app-cli-cca-creation',
+      properties,
+      timestamp: new Date(),
+      anonymousId: 'anonymous' 
+    });
     // eslint-disable-next-line no-empty
   } catch (e) {
-    // we just want to ignore an error from the tracking service
+    // ignore any error, to not block the cca run
   }
 }
